shamu: introduce a much less crippled liblights
Each color (RGB) of the tri-color LED sits on it's own GPIO, which
is a rather unique design. As such, the leds-qpnp driver is not
currently developed or designed for robust LED controls. Kernel
dtsi LED configuration will need to be overhauled for a fully
comprehensive solution.

This hal introduces basic user controls including custom charge and
notification colors. Blink and breathe are not supported at this time.

Change-Id: Ica9bc2aa820d1f7182e56aea54a66f7d1f9cde79

	modified:   init.shamu.rc
	modified:   liblight/Android.mk
	new file:   liblight/MODULE_LICENSE_APACHE2
	new file:   liblight/NOTICE
	new file:   liblight/lights.c
	deleted:    liblight/lights.shamu.so
	modified:   overlay/frameworks/base/core/res/res/values/config.xml